update crud employee

+++ src/main/java/vn/edu/stu/backend_service/controller/response/PageResponseAbstract.java
@@ -9,8 +9,8 @@ import lombok.Setter;
 @Setter
 
 public class PageResponseAbstract {
-    private int pageNumber;
-    private int pageSize;
-    private int totalPages;
-    private int totalElements;
+    private long pageNumber;
+    private long pageSize;
+    private long totalPages;
+    private long totalElements;
 }
